consolidate snippets file
